+++ deepforest/__init__.py
@@ -1,4 +1,3 @@
-from .scanner import ImageScanner
 from .cascade import CascadeForestClassifier
 from .forest import RandomForestClassifier
 from .forest import ExtraTreesClassifier
@@ -7,7 +6,6 @@ from .tree import ExtraTreeClassifier
 
 
 __all__ = [
-    "ImageScanner",
     "CascadeForestClassifier",
     "RandomForestClassifier",
     "ExtraTreesClassifier",
