+++ docs/api_reference.rst
@@ -3,16 +3,6 @@ API Reference
 
 Below is the class and function reference for :mod:`deepforest`. Notice that the package is under active development, and some features may not be stable yet.
 
-ImageScanner
-------------
-
-.. autoclass:: deepforest.ImageScanner
-    :members:
-    :inherited-members:
-    :show-inheritance:
-    :no-undoc-members:
-    :member-order: bysource
-
 CascadeForestClassifier
 -----------------------
